added components file, weatherScreen file and installed react-router-dom
